+++ 2D_Graphics_Project/Group_Project/Ellipse.h
@@ -22,8 +22,8 @@ class Ellipse: public Shape
 
         void Draw() override;
         void Move(int x, int y) override;
-        int Perimeter() override;
-        float Area() override;
+        double Perimeter() override;
+        double Area() override;
     
     private:
         int majorAxis
